QJniHelpers: impove error reporting (1)

See merge request !63

+++ QtAndroidHelpers/QAndroidDisplayMetrics.h
@@ -13,13 +13,13 @@
   modification, are permitted provided that the following conditions are met:
 
   * Redistributions of source code must retain the above copyright notice,
-    this list of conditions and the following disclaimer.
+	this list of conditions and the following disclaimer.
   * Redistributions in binary form must reproduce the above copyright notice,
-    this list of conditions and the following disclaimer in the documentation
-    and/or other materials provided with the distribution.
+	this list of conditions and the following disclaimer in the documentation
+	and/or other materials provided with the distribution.
   * Neither the name of the DoubleGIS, LLC nor the names of its contributors
-    may be used to endorse or promote products derived from this software
-    without specific prior written permission.
+	may be used to endorse or promote products derived from this software
+	without specific prior written permission.
 
   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
   AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
@@ -74,7 +74,7 @@ private:
 	Q_PROPERTY(float density READ density)
 	Q_PROPERTY(int densityDpi READ densityDpi)
 	Q_PROPERTY(float scaledDensity READ scaledDensity)
-	Q_PROPERTY(int densityFromDpi READ densityFromDpi)
+	Q_PROPERTY(float densityFromDpi READ densityFromDpi)
 	Q_PROPERTY(float scaledDensityFromDpi READ scaledDensityFromDpi)
 	Q_PROPERTY(float xdpi READ xdpi)
 	Q_PROPERTY(float ydpi READ ydpi)
